# TST Speed up some of the slowest tests