Mostly by reducing the dataset sizes.

- slowest tests for `pylatest_conda_forge_mkl` on `main`:

```
============================= slowest 20 durations =============================
31.32s call     decomposition/_dict_learning.py::sklearn.decomposition._dict_learning.DictionaryLearning
26.73s call     ensemble/tests/test_gradient_boosting.py::test_classification_synthetic[5-log_loss]
25.88s call     neighbors/tests/test_neighbors.py::test_knn_forcing_backend[kd_tree-multiprocessing]
25.35s call     neighbors/tests/test_neighbors.py::test_knn_forcing_backend[auto-multiprocessing]
23.85s call     ensemble/tests/test_gradient_boosting.py::test_classification_synthetic[5-exponential]
22.09s call     neighbors/tests/test_neighbors.py::test_knn_forcing_backend[ball_tree-multiprocessing]
18.64s call     tree/tests/test_tree.py::test_min_impurity_decrease
17.27s call     decomposition/tests/test_sparse_pca.py::test_mini_batch_correct_shapes
14.74s call     experimental/tests/test_enable_successive_halving.py::test_imports_strategies
14.27s call     utils/tests/test_estimator_checks.py::test_check_estimator_clones
13.12s call     experimental/tests/test_enable_iterative_imputer.py::test_imports_strategies
12.71s call     decomposition/tests/test_dict_learning.py::test_cd_work_on_joblib_memmapped_data
11.94s call     utils/tests/test_parallel.py::test_dispatch_config_parallel[2]
11.44s call     feature_selection/tests/test_sequential.py::test_n_features_to_select_stopping_criterion[forward]
11.35s call     ensemble/tests/test_gradient_boosting.py::test_gradient_boosting_validation_fraction
11.28s call     decomposition/tests/test_sparse_pca.py::test_transform_inverse_transform_round_trip[MiniBatchSparsePCA]
10.55s call     decomposition/tests/test_dict_learning.py::test_sparse_encode_shapes_omp
10.18s call     feature_selection/tests/test_rfe.py::test_rfe_cv_groups
9.86s call     feature_extraction/image.py::sklearn.feature_extraction.image.PatchExtractor
9.83s call     preprocessing/tests/test_target_encoder.py::test_fit_transform_not_associated_with_y_if_ordinal_categorical_is_not[5]
```

- same report on this PR:

```
============================= slowest 20 durations =============================
13.73s call     utils/tests/test_parallel.py::test_dispatch_config_parallel[2]
11.97s call     experimental/tests/test_enable_successive_halving.py::test_imports_strategies
11.21s call     decomposition/tests/test_dict_learning.py::test_cd_work_on_joblib_memmapped_data
11.06s call     model_selection/tests/test_split.py::test_nested_cv
10.62s call     decomposition/tests/test_sparse_pca.py::test_transform_inverse_transform_round_trip[MiniBatchSparsePCA]
10.44s call     decomposition/tests/test_dict_learning.py::test_dict_learning_lassocd_readonly_data
10.43s call     feature_extraction/image.py::sklearn.feature_extraction.image.PatchExtractor
10.37s call     feature_selection/tests/test_sequential.py::test_n_features_to_select_stopping_criterion[forward]
10.32s call     experimental/tests/test_enable_iterative_imputer.py::test_imports_strategies
9.76s call     decomposition/tests/test_dict_learning.py::test_sparse_encode_shapes_omp
9.11s call     metrics/tests/test_pairwise.py::test_sparse_manhattan_readonly_dataset
9.04s call     linear_model/tests/test_sgd.py::test_multi_core_gridsearch_and_early_stopping
8.89s call     ensemble/tests/test_gradient_boosting.py::test_gradient_boosting_validation_fraction
8.79s call     utils/tests/test_estimator_checks.py::test_check_estimator_clones
8.73s call     inspection/tests/test_partial_dependence.py::test_partial_dependence_non_null_weight_idx[1-estimator2]
8.69s call     preprocessing/tests/test_target_encoder.py::test_fit_transform_not_associated_with_y_if_ordinal_categorical_is_not[5]
8.67s call     inspection/tests/test_partial_dependence.py::test_partial_dependence_non_null_weight_idx[0-estimator2]
8.55s call     model_selection/tests/test_search.py::test_random_search_bad_cv
8.34s call     manifold/tests/test_t_sne.py::test_tsne_with_mahalanobis_distance
8.10s call     inspection/tests/test_partial_dependence.py::test_partial_dependence_equivalence_equal_sample_weight[LogisticRegression-data1]
```