@ogrisel 2023-09-15
> I am wondering if we could have been slightly more lenient with the test: be sure that fit/predict/transform works but not always running statistical checking for all tests.

In the long term, we will remove the tests for scipy sparse matrices and only keep the scipy sparse arrays.

In the short term, we could maybe change the defintion of the `*_CONTAINERS` lists to:

- only include scipy sparse matrices for versions of scipy < 1.8 (already the case)
- only include scipy sparse arrays for recent versions of scipy
- run both only if a specific environment variable is set and we would enable this on one of the fast runners with a recent version of scipy only.