@ogrisel 2023-09-16
Merged. There are more tests to optimize down the line but with diminishing returns.

Another PR that should greatly speed-up the tests is the test refactoring of pairwise argkmin & radius reductions: #27281.